// react-app/src/tutorial.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Command =
  | {
      checkIfFilesExist?: string[];
    }
  | {
      terminalCommands?: string[];
    }
  | {
      openFile?: string;
    }
  | {
      automaticImport?: {
        imports?: string[];
        path?: string[];
        [k: string]: unknown;
      };
    }
  | {
      fileDiff?: {
        filename?: string;
        solution?: string;
        [k: string]: unknown;
      };
    };
export type Instruction = Html | Image | Hint | CommandButton;

export interface Tutorial {
  title: string;
  description: string;
  tutorialFolder: string;
  exercises?: Exercise[];
}
export interface Exercise {
  title: string;
  description: string;
  checkStartState?: Command[];
  buildExercise?: Command[];
  cleanState?: Command[];
  test?: Command[];
  solve?: Command[];
  content?: Instruction[];
}
export interface Html {
  html: string;
}
export interface Image {
  image: {
    src: string;
    description?: string;
    width?: string;
  };
}
export interface Hint {
  hint: {
    text?: string;
    content?: Instruction[];
  };
}
export interface CommandButton {
  button: {
    text?: string;
    commands?: Command[];
  };
}